begin interaction with api

+++ lib/wow/commands/init.rb
@@ -1,4 +1,3 @@
-
 class Wow::Command::Init
   def initialize
 
@@ -7,6 +6,12 @@ class Wow::Command::Init
   def run
     src = Wow::Config.template_path('packages.toml')
     dst = File.join(Dir.pwd, Wow::Package::Config.filename)
-    FileUtils.cp(src, dst)
+    if File.exist? dst
+      unless agree("#{Wow::Package::Config.filename} already exists in this folder are you sure you want to override it? [yn]")
+        return
+      end
+      FileUtils.cp(src, dst)
+      puts "Created config successfully in #{Wow::Package::Config.filename}"
+    end
   end
 end
